added email verification

+++ server.js
@@ -1,19 +1,19 @@
 const express = require("express")
 const mongoose = require("mongoose")
-const { Router } = require("express")
-const router = Router()
 const app = express()
 const bodyParser = require("body-parser")
 const cors = require("cors")
 const Schema = mongoose.Schema
+const nodemailer = require("nodemailer")
+const crypto = require("crypto")
 const bcrypt = require("bcryptjs")
-require("dotenv/config")
+require("dotenv").config({ path: "./.env" })
 
 const port = process.env.PORT || 8020
 
 // --- MongoDB connection ---
 
-// const uri = process.env.PROD_MONGODB || process.env.MONGODB_URI
+const uri = process.env.PROD_MONGODB || process.env.MONGODB_URI
 // create a mongoDB new connection
 
 const options = {}
@@ -27,9 +27,11 @@ const db = mongoose.connection
 
 db.on("error", console.error.bind(console, "MongoDB connection error:"))
 
-const Account = new Schema({
+const AccountDB = new Schema({
   admin: { type: Boolean, default: false },
   subscription: { type: String, default: "free" },
+  verified: { type: Boolean, default: false },
+  verificationToken: { type: String },
   name: { type: String, required: true },
   email: { type: String, required: true },
   username: { type: String, required: true },
@@ -37,6 +39,21 @@ const Account = new Schema({
   createdAt: { type: Date, default: Date.now },
   updatedAt: { type: Date, default: Date.now },
 })
+
+const Account = mongoose.model("Account", AccountDB)
+
+const emailValidation =
+  /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+
+/** Password validation
+ * (?=.*\d)         should contain at least 1 digit
+ * (?=(.*\W){2})    should contain at least 2 special characters
+ * (?=.*[a-zA-Z])   should contain at least 1 alphabetic character
+ * (?!.*\s)         should not contain any blank space
+ */
+
+const passwordValidation =
+  /^(?=.*\d)(?=(.*\W){2})(?=.*[a-zA-Z])(?!.*\s).{1,15}$/
 
 // --- MIDDLEWARE ---
 
@@ -61,23 +78,129 @@ const handleCallback = (callback) => async (req, res, next) => {
 // --- CONTROLLERS ---
 
 const createAccount = handleCallback(async (req, res) => {
-  const salt = await bcrypt.genSalt(10)
-  const hashedPassword = await bcrypt.hash(req.body.password, salt)
+  if (
+    !req.body.username ||
+    !req.body.email ||
+    !req.body.password ||
+    !req.body ||
+    !req.body.name
+  ) {
+    throw new Error("Missing username, name, email or password")
+  }
 
-  const account = await new Account({
-    username: req.body.username,
-    email: req.body.email,
-    password: hashedPassword,
-  })
+  if (req.body.password.length < 6) {
+    throw new Error("Password must be at least 6 characters")
+  }
 
-  await account.save()
-  return res.status(201).json({
-    success: true,
-    message: "Account created",
-  })
+  if (!emailValidation.test(req.body.email)) {
+    throw new Error("Invalid email")
+  }
+
+  if (!passwordValidation.test(req.body.password)) {
+    throw new Error("Invalid password")
+  }
+
+  const doesAccoutExist = await Account.findOne({ email: req.body.email })
+  if (doesAccoutExist) {
+    throw new Error("Account already exists")
+  }
+
+  if (!doesAccoutExist) {
+    const salt = await bcrypt.genSalt(10)
+    const hashedEmail = await bcrypt.hash(req.body.email, salt)
+    const hashedPassword = await bcrypt.hash(req.body.password, salt)
+    const verificationToken = await crypto.randomBytes(20).toString("hex")
+
+    const account = await new Account({
+      name: req.body.name,
+      username: req.body.username,
+      email: hashedEmail,
+      password: hashedPassword,
+      verificationToken: verificationToken,
+    })
+
+    await account.save()
+
+    sendVerificationEmail({ email: req.body.email, verificationToken }, res)
+
+    return res.status(201).json({
+      success: true,
+      message: "Account created",
+    })
+  }
 })
 
+const sendVerificationEmail = async (req, res) => {
+  const smtpConfig = {
+    host: "smtp.zoho.com",
+    port: 465,
+    secure: true, //ssl
+    auth: {
+      user: process.env.EMAIL,
+      pass: process.env.PASSWORD,
+    },
+  }
+
+  const transporter = nodemailer.createTransport(smtpConfig)
+
+  const mailOptions = {
+    from: process.env.EMAIL,
+    to: req.email,
+    subject: "Account verification",
+    text: `Click on the link below to verify your account: ${process.env.BASE_URL}/verify?token=${req.verificationToken}`,
+  }
+
+  transporter.sendMail(mailOptions, (error, info) => {
+    if (error) {
+      console.log(error)
+    } else {
+      console.log("Email sent: " + info.response)
+    }
+  })
+  return res.status(200).json({ success: true, message: "Email sent" })
+}
+
+const verify = async (req, res) => {
+  const { token } = req.query
+
+  if (!token) {
+    throw new Error("Missing token")
+  }
+
+  const updatedAccount = await Account.findOneAndUpdate(
+    { verificationToken: token },
+    { verified: true, verificationToken: undefined }
+  )
+
+  if (!updatedAccount) {
+    throw new Error("Invalid token")
+  }
+
+  if (!updatedAccount.verified) {
+    throw new Error("Account not verified")
+  }
+
+  updatedAccount.email = undefined
+  updatedAccount.password = undefined
+
+  return res.status(200).json({ success: true, data: updatedAccount })
+}
+
 const signin = async (req, res) => {
+  let activeAccount = null
+
+  if (!req.body.username || !req.body.password) {
+    throw new Error("Missing username or password")
+  }
+
+  if (!emailValidation.test(req.body.email)) {
+    throw new Error("Invalid email")
+  }
+
+  if (!passwordValidation.test(req.body.password)) {
+    throw new Error("Invalid password")
+  }
+
   activeAccount = await Account.findOne({ username: req.body.username })
   !activeAccount && res.status(400).json("Wrong credentials")
 
@@ -87,55 +210,61 @@ const signin = async (req, res) => {
   )
   !validPassword && res.status(400).json("Wrong credentials")
 
+  activeAccount.email = undefined
   activeAccount.password = undefined
 
   res.status(200).json({ success: true, data: activeAccount })
 }
 
-const verify = async (req, res) => {
-  const verified = await Account.validate({
-    email: req.body.email,
-    password: req.body.password,
-  })
-  if (!verified)
-    throw {
-      success: false,
-      message: "Wrong credentials",
-    }
-  return res.status(200).json({ success: true, data: verified })
-}
-
-const updateAccount = handleCallback(async ({ body, params }, res) => {
-  const { id } = params
-
-  if (req.body.id === id) {
+const updateAccount = handleCallback(async (req, res, next) => {
+  if (req.body.id === req.params.id || req.body.isAdmin) {
     if (req.body.password) {
       const salt = await bcrypt.genSalt(10)
-      body.password = await bcrypt.hash(body.password, salt)
+      req.body.password = await bcrypt.hash(req.body.password, salt)
     }
-    const account = await Account.findByIdAndUpdate(id, { $set: req.body })
-    return res.status(200).json({ success: true, message: "Account updated" })
-  } else {
-    throw new Error("Unauthorized")
+
+    await Account.findByIdAndUpdate(req.params.id, {
+      $set: req.body,
+    })
+
+    res.status(200).json("Account has been updated")
+  } else if (!req.body.isAdmin) {
+    const account = await Account.findById(req.params.id)
+    if (account) {
+      return res.status(403).json("you can only update your account!")
+    }
   }
+  throw new Error("Account not found")
 })
 
 const deleteAccount = handleCallback(async (req, res) => {
   const { id } = req.params
-  const deleted = await Account.findByIdAndDelete(id)
-  if (deleted) {
-    return res
-      .status(200)
-      .json({ success: true, data: deleted, Message: "Account deleted" })
+  if (!id) throw new Error("Account not found")
+
+  if (req.body.id === req.params.id || req.body.isAdmin) {
+    const account = await Account.findById(id)
+    if (!account) throw new Error("Account not found")
+
+    const deleted = await Account.findByIdAndDelete(id)
+    if (!deleted) throw new Error("Account not found")
+
+    if (deleted) {
+      deleted.password = undefined
+      deleted.email = undefined
+      return res
+        .status(200)
+        .json({ success: true, data: deleted, Message: "Account deleted" })
+    }
   }
   throw new Error("Account not found")
 })
 
 const getAccount = handleCallback(async (req, res) => {
-  console.log(req.params)
-  const { id } = req.params
-  const account = await Account.findById(id)
-  if (account) {
+  if (!req.params.id) throw new Error("Account not found")
+
+  if (req.body.id === req.params.id && req.body.isAdmin) {
+    const account = await Account.findById(req.params.id)
+    if (!account) throw new Error("Account not found")
     return res.status(200).json({ success: true, data: account })
   }
   throw new Error("Account not found")
@@ -143,13 +272,15 @@ const getAccount = handleCallback(async (req, res) => {
 
 // --- ROUTES ---
 
-router.get("/", (res, req) => res.json("base api endpoint"))
-router.post("/signin", signin)
-router.post("/verify", verify)
-router.get("/account/:id", getAccount)
-router.put("/account/:id", updateAccount)
-router.post("/account/create", createAccount)
-router.delete("/account/:id", deleteAccount)
+app.get("/", (req, res) => res.json("base api endpoint"))
+app.post("/signin", signin)
+app.get("/verify/:token", verify)
+app.post("/account/create", createAccount)
+app.get("/account/:id", getAccount)
+app.put("/account/:id", updateAccount)
+app.delete("/account/:id", deleteAccount)
+
+// --- SERVER ---
 
 app.listen(port, () => {
   console.log(`Listening at http://localhost:${port}`)
